+++ src/SqlChangeTracker/Config/Constants.cs
@@ -12,7 +12,6 @@ internal static class ExitCodes
 internal static class ConfigFileNames
 {
     public const string SqlctConfigFileName = "sqlct.config.yaml";
-    public const string SqlctConfigLegacyFileName = "sqlct.config.json";
 }
 
 internal static class ErrorCodes